Move `OfflineDocumentGateway` to a shared location
